+++ tools/delete_pubspec_overrides.dart
@@ -1,6 +1,0 @@
-import 'dart:io';
-
-void main(List<String> arguments) => Directory.current
-    .list(recursive: true)
-    .where((f) => f.path.endsWith('pubspec_overrides.yaml'))
-    .forEach((f) => f.deleteSync());